requirements.txt
streamlit==1.29.0
crewai==0.28.8
langchain==0.1.5
langchain-community==0.0.13
ollama==0.1.7
duckduckgo-search==4.1.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.25.2
plotly==5.17.0
markdown==3.5.2
